test: add test cases for keyboard operations
