+++ examples/mutiple-with-maxCount.tsx
@@ -46,8 +46,8 @@ export default () => {
         multiple
         maxCount={3}
         treeData={treeData}
-        onChange={onChange}
-        value={value}
+        // onChange={onChange}
+        // value={value}
       />
 
       <h2>checkable with maxCount</h2>
@@ -55,10 +55,21 @@ export default () => {
         style={{ width: 300 }}
         multiple
         treeCheckable
+        treeCheckStrictly
         maxCount={3}
         treeData={treeData}
         onChange={onChange}
         value={value}
+      />
+      <TreeSelect
+        style={{ width: 300 }}
+        treeData={[
+          { key: '0', value: '0', title: '0 label' },
+          { key: '1', value: '1', title: '1 label' },
+          { key: '2', value: '2', title: '2 label' },
+        ]}
+        multiple
+        maxCount={2}
       />
     </>
   );
